+++ example/android/app/src/main/kotlin/com/zellar/zcomponents/z_components_example/MainActivity.kt
@@ -1,13 +1,7 @@
 package com.zellar.zcomponents.z_components_example
 
-import android.os.Bundle
-
-import io.flutter.app.FlutterActivity
-import io.flutter.plugins.GeneratedPluginRegistrant
+import io.flutter.embedding.android.FlutterActivity
 
 class MainActivity: FlutterActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    GeneratedPluginRegistrant.registerWith(this)
-  }
+
 }