+++ ios/Classes/ZComponentsPlugin.m
@@ -1,8 +1,0 @@
-#import "ZComponentsPlugin.h"
-#import <z_components/z_components-Swift.h>
-
-@implementation ZComponentsPlugin
-+ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [SwiftZComponentsPlugin registerWithRegistrar:registrar];
-}
-@end