+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.9.41
+
+* Atualizar para o flutter 2.8
+
+## 0.9.40+12
+
+* Ajuste Luiz
+
+## 0.9.40+11
+
+* Ajuste ZDocumentos seta na lista de documentos
+
+## 0.9.40+10
+
+* Ajuste ZDocumentos
+
 ## 0.9.40+9
 
 * Ajuste ZCollection